gpu-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
redis==5.0.1
boto3==1.29.7
cryptography==41.0.7
spacy==3.7.2
whisper-openai==20231117
face-recognition==1.3.0
pytesseract==0.3.10
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
prometheus-client==0.19.0
psutil==5.9.6
pytest==7.4.3
pytest-asyncio==0.21.1